Add user profile editing
Only display action buttons for user/pushup record owner
Prevent indirect edit access to user/pushup records

+++ resources/views/pushups/show.blade.php
@@ -12,12 +12,14 @@
                     <h4 class>Record Detail</h4>
                 </div>
                 
-                <ul class="list-group list-group-flush">
-                    <li class="list-group-item grid-bg">
-                        <a class="btn btn-outline-primary btn-sm " href="/pushups/{{ $pushup->id }}/edit">Edit</a>
-                        <a class="btn btn-outline-primary btn-sm " href="/pushups/delete">Delete</a>
-                    </li>
-                </ul>
+                @if (auth()->id() == $pushup->user->id)
+                    <ul class="list-group list-group-flush">
+                        <li class="list-group-item grid-bg">
+                            <a class="btn btn-outline-primary btn-sm " href="/pushups/{{ $pushup->id }}/edit">Edit</a>
+                            <a class="btn btn-outline-primary btn-sm " href="/pushups/delete">Delete</a>
+                        </li>
+                    </ul>
+                @endif
 
                 <div class="card-body">
